refactor git-secret #11
